+++ moco-doc/ReleaseNotes.md
@@ -1,3 +1,8 @@
+# Release
+
+## APIs
+* Latency with time unit
+
 # Release 0.10.0 (1-Dec-2014)
 
 ## Usage